need to fix: filters apply to first page only and products in cart dont remain if the page refreshes

+++ main/delete.php
@@ -3,7 +3,7 @@
 session_start();
 $mysqli = require __DIR__ . "/conn.php";
 
-if (isset($_POST["deleteBtn"])) {
+if ($_SERVER["REQUEST_METHOD"] == "POST" && isset($_POST["deleteBtn"])) {
     $stmt = $mysqli->prepare("DELETE FROM ram WHERE product_id = ?");
     $stmt->bind_param("i", $_POST['product_id']);
